allow running as a different user
Introducing a -user flag that accepts either a POSIX uid or user name.
This will be used to run the user's application (i.e. the subprocess)

This can be used in the Dockerfile as

    RUN adduser --disabled-password foo
    ENTRYPOINT ["/runsd", "-user=foo", "--", "/myapp"]

or alternatively with a predefined uid:

    RUN adduser --disabled-password --uid 1234 foo
    ENTRYPOINT ["/runsd", "-user=1234", "--", "/myapp"]

Might add support for gid/group name later.
Also adding /uid endpoint to the ./example app to help users try this out.

Resolves #16.

Signed-off-by: Ahmet Alp Balkan <ahmetb@google.com>